Added yunCUL OLED println options: ++ means turn on OLED display -- means turn off OLED display
This provides a way to create _some_kind_of OLED-screen-saver in FHEM.
FHEM is so powerful.. It must be capable of turning off the OLED after a timeout has expired.

Every time OLED println() is issued, the dislay s turned on.
FHEM should take care of disabling the OLED-screen after a minute (or so).